# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Artifact Registry Backend",
    "version": "0.1.9"
  },
  "paths": {
...
    "/api/v1/events/{artifact_type}/{artifact_id}/transition": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Transition Artifact",
        "operationId": "transition_artifact_api_v1_events__artifact_type___artifact_id__transition_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Type"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusTransition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{artifact_type}/{artifact_id}/history": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Artifact History",
        "operationId": "get_artifact_history_api_v1_events__artifact_type___artifact_id__history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Type"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactEventOut"
                  },
                  "title": "Response Get Artifact History Api V1 Events  Artifact Type   Artifact Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/comments/": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "List Comments",
        "description": "List all comments for an artifact.\nOptionally filter by resolved status.",
        "operationId": "list_comments_api_v1_comments__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "artifact_aid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Artifact AID to get comments for",
              "title": "Artifact Aid"
            },
            "description": "Artifact AID to get comments for"
          },
          {
            "name": "resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resolution status",
              "title": "Resolved"
            },
            "description": "Filter by resolution status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  },
                  "title": "Response List Comments Api V1 Comments  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "Create Comment",
        "description": "Create a new comment on an artifact field.",
        "operationId": "create_comment_api_v1_comments__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comments/{comment_id}/resolve": {
      "patch": {
        "tags": [
          "comments"
        ],
        "summary": "Resolve Comment",
        "description": "Mark a comment as resolved.",
        "operationId": "resolve_comment_api_v1_comments__comment_id__resolve_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentResolve"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comments/{comment_id}/unresolve": {
      "patch": {
        "tags": [
          "comments"
        ],
        "summary": "Unresolve Comment",
        "description": "Mark a comment as unresolved (reopen it).",
        "operationId": "unresolve_comment_api_v1_comments__comment_id__unresolve_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/comments/{comment_id}": {
      "delete": {
        "tags": [
          "comments"
        ],
        "summary": "Delete Comment",
        "description": "Delete a comment.",
        "operationId": "delete_comment_api_v1_comments__comment_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactEventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "artifact_type": {
            "type": "string",
            "title": "Artifact Type"
          },
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "event_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Event Data"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "id",
          "artifact_type",
          "artifact_id",
          "event_type",
          "event_data",
          "timestamp"
        ],
        "title": "ArtifactEventOut"
      },
...
      "Comment": {
        "properties": {
          "artifact_aid": {
            "type": "string",
            "title": "Artifact Aid"
          },
          "field_name": {
            "type": "string",
            "title": "Field Name"
          },
          "comment_text": {
            "type": "string",
            "title": "Comment Text"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "resolved": {
            "type": "boolean",
            "title": "Resolved"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "selected_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Text"
          },
          "resolution_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Action"
          }
        },
        "type": "object",
        "required": [
          "artifact_aid",
          "field_name",
          "comment_text",
          "author",
          "id",
          "created_at",
          "resolved"
        ],
        "title": "Comment"
      },
      "CommentCreate": {
        "properties": {
          "artifact_aid": {
            "type": "string",
            "title": "Artifact Aid"
          },
          "field_name": {
            "type": "string",
            "title": "Field Name"
          },
          "comment_text": {
            "type": "string",
            "title": "Comment Text"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "selected_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Text"
          }
        },
        "type": "object",
        "required": [
          "artifact_aid",
          "field_name",
          "comment_text",
          "author"
        ],
        "title": "CommentCreate"
      },
      "CommentResolve": {
        "properties": {
          "resolved_by": {
            "type": "string",
            "title": "Resolved By"
          },
          "resolution_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Action"
          }
        },
        "type": "object",
        "required": [
          "resolved_by"
        ],
        "title": "CommentResolve"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "Draft",
          "Ready_for_Review",
          "In_Review",
          "Approved",
          "Deferred",
          "Rejected",
          "Superseded",
          "Retired"
        ],
        "title": "Status"
      },
      "StatusTransition": {
        "properties": {
          "from_status": {
            "$ref": "#/components/schemas/Status"
          },
          "to_status": {
            "$ref": "#/components/schemas/Status"
          },
          "rationale": {
            "type": "string",
            "minLength": 1,
            "title": "Rationale"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "from_status",
          "to_status",
          "rationale"
        ],
        "title": "StatusTransition"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/token"
          }
        }
      }
    }
  }
}